refactor: replace generic transport parameter with mock toggle
- Abstract Transport trait is now a concrete implementation with #[cfg] toggles
- Use MockTransport for tests via compile-time switching
- Remove Libp2pTransport abstraction and flatten directory structure
- Move libp2p_transport module content to top-level transport
- Fix struct definitions to remove generic type parameters
